+++ package.json
@@ -2,6 +2,7 @@
   "name": "skoleblogpwa2",
   "version": "0.1.0",
   "private": true,
+  "homepage": "http://jtad009.github.io/SkoleBlogReactClient",
   "dependencies": {
     "react": "^16.9.0",
     "react-dom": "^16.9.0",
@@ -13,7 +14,9 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "predeploy": "npm run build",
+    "deploy": "gh-pages -d build"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -29,5 +32,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "gh-pages": "^2.1.1"
   }
 }
